added infinity

+++ rainmaker/pt-calculator-v2/src/main/java/org/egov/pt/calculator/repository/rowmapper/BillingSlabRowMapper.java
@@ -27,9 +27,9 @@ public class BillingSlabRowMapper implements ResultSetExtractor<List<BillingSlab
 						.lastModifiedBy(rs.getString("lastmodifiedby")).lastModifiedTime(rs.getLong("lastmodifiedtime")).build();
 				
 				currentBillingSlab = BillingSlab.builder().id(rs.getString("id")).areaType(rs.getString("areaType"))
-						.fromFloor(rs.getInt("fromFloor")).fromPlotSize(rs.getDouble("fromPlotSize")).ownerShipCategory(rs.getString("ownerShipCategory"))
+						.fromFloor(rs.getDouble("fromFloor")).fromPlotSize(rs.getDouble("fromPlotSize")).ownerShipCategory(rs.getString("ownerShipCategory"))
 						.propertySubType(rs.getString("propertySubType")).propertyType(rs.getString("propertyType")).subOwnerShipCategory(rs.getString("subOwnerShipCategory"))
-						.tenantId(rs.getString("tenantId")).toFloor(rs.getInt("toFLoor")).toPlotSize(rs.getDouble("toPlotSize")).unitRate(rs.getDouble("unitRate"))
+						.tenantId(rs.getString("tenantId")).toFloor(rs.getDouble("toFLoor")).toPlotSize(rs.getDouble("toPlotSize")).unitRate(rs.getDouble("unitRate"))
 						.usageCategoryDetail(rs.getString("usageCategoryDetail")).usageCategoryMajor(rs.getString("usageCategoryMajor")).usageCategoryMinor(rs.getString("usageCategoryMinor"))
 						.usageCategorySubMinor(rs.getString("usageCategorySubMinor")).auditDetails(auditDetails).build();
